Adapt data types to Windows
Type "long" seems to be 32bit on Win, but 64bit on Linux

+++ src/support/http-handler.cpp
@@ -299,7 +299,7 @@ void HttpHandler::run()
     terminating();
     return;
   }
-  u32 mode = 1;
+  unsigned long mode = 1;
   ioctlsocket(ListenSocket, FIONBIO, &mode);
 
   // Setup the TCP listening socket
